requirements.txt
asgiref==3.6.0
crispy-bootstrap5==0.7
Django==4.1.7
django-autoslug==1.9.8
django-crispy-forms==2.0
Pillow==9.5.0
sqlparse==0.4.3
tzdata==2023.3
